Header check copies existing fields.
Specific tags already specified in the file could be copied into the
header. Some default values can be made up in other instances (i.e.
date).

The old header (if any) is maintained due to it possibly containing
information which should not be discarded... and the point of figuring
out whether it is actually the header.